backend/requirements.txt
Django==5.2.1
django-allauth[mfa,socialaccount]>=65.8.0
qrcode >= 7.0.0
django-sendgrid-v5==1.2.3
dj-stripe==2.8.4
stripe
django-unfold==0.30.0
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.6
gunicorn==20.1.0
whitenoise==6.7.0
fido2==1.1.3